Changed Order, Currency implementation
