+++ src/MuHub.Market.Proxy/Features/Coins/Currency.cs
@@ -1,7 +1,14 @@
 ﻿namespace MuHub.Market.Proxy.Features.Coins;
 
-public enum Currency
+public sealed class Currency
 {
-    Usd,
-    Eur,
+    public static readonly Currency Usd = new(CoinGecko.Api.Features.Coins.VsCurrency.Usd);
+    public static readonly Currency Eur = new(CoinGecko.Api.Features.Coins.VsCurrency.Eur);
+
+    private Currency(CoinGecko.Api.Features.Coins.VsCurrency value)
+    {
+        Value = value;
+    }
+
+    internal CoinGecko.Api.Features.Coins.VsCurrency Value { get; }
 }
